Add duplicate advancements feature
Adds a config called allow-duplicate-advancements that determines whether or not to count duplicate advancements between players on the server.
E.g. if two players have the advancement 'Stone Age' it will be counted both times, but with this option off it would only count as one.
This is on by default so behavior doesn't change for current users of the plugin.